fix(build): make Spring sample shadowJar patching config-cache safe
